Resources/python/sensors/requirements.txt
paho-mqtt==1.6.1
python-dotenv==1.0.0
dataclasses-json==0.6.1
structlog==24.1.0